validator_api/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

Django<1.9,>=1.8 # BSD
djangorestframework == 3.3.3
django-cors-headers == 1.1.0

# User Authentication
# -e git+https://github.com/ging/django-openstack-auth.git#egg=django-openstack-auth
# -e git+https://github.com/ging/python-keystoneclient.git#egg=python-keystoneclient
pbr>=1.6 # Apache-2.0
cryptography
oslo.config>=3.9.0 # Apache-2.0
oslo.policy>=0.5.0 # Apache-2.0
oslo.log >= 1.8.0
python-keystoneclient!=1.8.0,!=2.1.0,>=1.7.0 # Apache-2.0
keystoneauth1>=2.1.0 # Apache-2.0
six>=1.9.0 # MIT

# Version Control
svn == 0.3.36
GitPython == 2.0.2

# Image Management
docker-py == 1.8.1